refactor: move internal types under own subfolders
* move deploy-loader types under own folder
* move journal types under own folder
* move execution types under own folder

Resolves #374.
