+++ packages/core/src/new-api/internal/deployer.ts
@@ -20,23 +20,23 @@ import {
 
 import { Batcher } from "./batcher";
 import { defaultConfig } from "./defaultConfig";
+import { DeploymentLoader } from "./deployment-loader/types";
 import { ExecutionEngine } from "./execution/execution-engine";
 import { executionStateReducer } from "./execution/execution-state-reducer";
 import { BasicExecutionStrategy } from "./execution/execution-strategy";
 import { TranactionLookupTimerImpl } from "./execution/transaction-lookup-timer";
+import {
+  ChainDispatcher,
+  ContractAtExecutionState,
+  DeploymentExecutionState,
+  ExecutionStateMap,
+  ExecutionStrategy,
+  TransactionLookupTimer,
+} from "./execution/types";
 import { ModuleConstructor } from "./module-builder";
 import { Reconciler } from "./reconciliation/reconciler";
 import { ArtifactMap } from "./reconciliation/types";
 import { isContractExecutionStateArray } from "./type-guards";
-import { ChainDispatcher } from "./types/chain-dispatcher";
-import { DeploymentLoader } from "./types/deployment-loader";
-import { ExecutionStrategy } from "./types/execution-engine";
-import {
-  ContractAtExecutionState,
-  DeploymentExecutionState,
-  ExecutionStateMap,
-} from "./types/execution-state";
-import { TransactionLookupTimer } from "./types/transaction-timer";
 import { assertIgnitionInvariant } from "./utils/assertions";
 import { getFuturesFromModule } from "./utils/get-futures-from-module";
 import { validate } from "./validation/validate";
