+++ packages/core/src/new-api/internal/reconciliation/futures/reconcileNamedLibraryDeployment.ts
@@ -1,7 +1,7 @@
 import { isEqual } from "lodash";
 
 import { NamedLibraryDeploymentFuture } from "../../../types/module";
-import { DeploymentExecutionState } from "../../types/execution-state";
+import { DeploymentExecutionState } from "../../execution/types";
 import { resolveFromAddress } from "../../utils/resolve-from-address";
 import { ExecutionStateResolver } from "../execution-state-resolver";
 import { ReconciliationContext, ReconciliationFutureResult } from "../types";
