+++ packages/core/src/new-api/internal/reconciliation/futures/reconcileNamedStaticCall.ts
@@ -1,7 +1,7 @@
 import { isEqual } from "lodash";
 
 import { NamedStaticCallFuture } from "../../../types/module";
-import { StaticCallExecutionState } from "../../types/execution-state";
+import { StaticCallExecutionState } from "../../execution/types";
 import { resolveFromAddress } from "../../utils/resolve-from-address";
 import { ExecutionStateResolver } from "../execution-state-resolver";
 import { ReconciliationContext, ReconciliationFutureResult } from "../types";
