+++ packages/core/src/new-api/internal/reconciliation/reconcile-dependency-rules.ts
@@ -5,7 +5,7 @@ import {
   ExecutionState,
   ExecutionStateMap,
   ExecutionStatus,
-} from "../types/execution-state";
+} from "../execution/types";
 
 import { ReconciliationFutureResult } from "./types";
 import { fail } from "./utils";
