+++ packages/core/test/new-api/reconciliation/futures/reconcileArtifactContractAt.ts
@@ -9,7 +9,7 @@ import {
   ExecutionStateMap,
   ExecutionStatus,
   StaticCallExecutionState,
-} from "../../../../src/new-api/internal/types/execution-state";
+} from "../../../../src/new-api/internal/execution/types";
 import { FutureType } from "../../../../src/new-api/types/module";
 import { exampleAccounts, initOnchainState } from "../../helpers";
 import { assertSuccessReconciliation, reconcile } from "../helpers";
