+++ packages/core/test/new-api/reconciliation/futures/reconcileArtifactContractDeployment.ts
@@ -6,7 +6,7 @@ import { buildModule } from "../../../../src/new-api/build-module";
 import {
   DeploymentExecutionState,
   ExecutionStatus,
-} from "../../../../src/new-api/internal/types/execution-state";
+} from "../../../../src/new-api/internal/execution/types";
 import { FutureType } from "../../../../src/new-api/types/module";
 import { exampleAccounts, initOnchainState } from "../../helpers";
 import {
