+++ packages/core/test/new-api/reconciliation/futures/reconcileReadEventArgument.ts
@@ -6,7 +6,7 @@ import {
   DeploymentExecutionState,
   ExecutionStatus,
   ReadEventArgumentExecutionState,
-} from "../../../../src/new-api/internal/types/execution-state";
+} from "../../../../src/new-api/internal/execution/types";
 import { FutureType } from "../../../../src/new-api/types/module";
 import { exampleAccounts, initOnchainState } from "../../helpers";
 import { assertSuccessReconciliation, reconcile } from "../helpers";
