+++ packages/core/test/new-api/reconciliation/futures/reconcileSendData.ts
@@ -5,7 +5,7 @@ import { buildModule } from "../../../../src/new-api/build-module";
 import {
   ExecutionStatus,
   SendDataExecutionState,
-} from "../../../../src/new-api/internal/types/execution-state";
+} from "../../../../src/new-api/internal/execution/types";
 import { FutureType } from "../../../../src/new-api/types/module";
 import { exampleAccounts, initOnchainState } from "../../helpers";
 import { assertSuccessReconciliation, reconcile } from "../helpers";
