+++ packages/core/test/new-api/reconciliation/reconciler.ts
@@ -2,11 +2,11 @@
 import { assert } from "chai";
 
 import { buildModule } from "../../../src/new-api/build-module";
-import { ArtifactMap } from "../../../src/new-api/internal/reconciliation/types";
 import {
   DeploymentExecutionState,
   ExecutionStatus,
-} from "../../../src/new-api/internal/types/execution-state";
+} from "../../../src/new-api/internal/execution/types";
+import { ArtifactMap } from "../../../src/new-api/internal/reconciliation/types";
 import { FutureType } from "../../../src/new-api/types/module";
 import { exampleAccounts, initOnchainState } from "../helpers";
